prevent multiple snitches from running
This change checks for already scheduled snitches before scheduling a new one, so if more than 1 snitch is running, they will be pruned down to 1.  To be more effective we could introduce a random delay incase multiple snitches were scheduled for the exact same time, but it seems to be working just fine for us as is.